Added data service to witan.ui
Data service is incomplete but it controls critical data. View-models
request 'data' (forecasts etc) from this service - NOT the api directly.
It has some transient global state as well such as `(logged-in?)` but
this should be used sparingly. Data service is not a dumping ground for
globals! All acquisition should go via requests.